Refactored ParcelCalendar to be a non-async component, with data now loading on server-side.
