+++ src/app/calendar/page.tsx
@@ -1,11 +1,14 @@
 import { Metadata } from "next";
 import React from "react";
 import ParcelCalendar from "@/app/calendar/ParcelCalendar";
+import getParcelCalendarEvents from "@/app/calendar/getParcelCalendarEvents";
 
-const CalendarPage: Promise<React.ReactElement> = async () => {
+const CalendarPage = async (): Promise<React.ReactElement> => {
+    const events = await getParcelCalendarEvents();
+
     return (
         <main>
-            <ParcelCalendar />
+            <ParcelCalendar events={events} />
         </main>
     );
 };
